refactor(library): own browse paging and cache in one route module
Move page-zero bootstrap, the normal-grid/virtual-mode decision,
random-access window fetching, per-page cache bridging, identity reset,
and failed-page retry out of the Library Browser route into one
route-owned module. Small libraries keep flowing through the bottom
sentinel; large libraries fetch only current-window pages plus the
bounded look-ahead in either display direction. Sort, filter, library,
or Saved Service Profile identity changes clear virtual state, reset
scroll, and reject stale completions; cached route re-entry reuses the
seeded page-zero and per-page entries; failed virtual pages stay
failures with retry over the current window. The layout and
page-selection utilities remain internal seams, and the route keeps
rendering and geometry.

Closes #207
